requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.1
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-extensions==2.1.2
django-rest-swagger==2.2.0
djangorestframework==3.8.2
fpdf==1.7.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2-binary==2.7.5
PyPDF2==1.26.0
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
